requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
twilio==9.2.4
channels==4.2.2
channels-redis==4.3.0
celery==5.4.0
redis==5.0.8
asgiref==3.8.1
sqlparse==0.5.2
cryptography==43.0.1
pytz==2024.2
qrcode==8.2
Pillow==11.3.0